+++ src/by/it/_tasks_/ivanstetsles1/HelloWorld.java
@@ -1,0 +1,7 @@
+package by.it._tasks_.ivanstetsles1;
+
+public class HelloWorld {
+    public static void main(String[] args) {
+        System.out.println("Hello World!");
+    }
+}
