+++ src/by/it/_tasks_/ivanstetsles2/TaskA1.java
@@ -1,4 +1,4 @@
-package by.it._tasks_.ivanstets;
+package by.it._tasks_.ivanstetsles2;
 /*
 напишите программу которая выводит на экран
 Hello world!
