+++ src/by/it/_tasks_/ivanstetsles2/TaskB1.java
@@ -1,4 +1,4 @@
-package by.it._tasks_.ivanstets;
+package by.it._tasks_.ivanstetsles2;
 
 /*
 Выводим квадрат числа
