+++ src/by/it/_tasks_/ivanstetsles2/TaskC2.java
@@ -1,4 +1,4 @@
-package by.it._tasks_.ivanstets;
+package by.it._tasks_.ivanstetsles2;
 
 import java.util.Scanner;
