+++ src/by/it/_tasks_/ivanstetsles3/TaskB2.java
@@ -1,4 +1,4 @@
-package by.it._tasks_.ivanstets2;
+package by.it._tasks_.ivanstetsles3;
 
 /*
 Lesson 03. Task B2. Нужно написать программу, которая вводит три числа  a b c типа int с клавиатуры
