+++ src/by/it/_tasks_/ivanstetsles3/TaskC2.java
@@ -1,4 +1,4 @@
-package by.it._tasks_.ivanstets2;
+package by.it._tasks_.ivanstetsles3;
 /*
     Сумма цифр четырехзначного числа
