+++ src/by/it/_tasks_/ivanstetsles4/TaskB2.java
@@ -1,4 +1,4 @@
-package by.it._tasks_.ivanstets3;
+package by.it._tasks_.ivanstetsles4;
 
 /*
 Таблица умножения
