+++ src/by/it/_tasks_/ivanstetsles4/TaskC1.java
@@ -1,4 +1,4 @@
-package by.it._tasks_.ivanstets3;
+package by.it._tasks_.ivanstetsles4;
 
 /*
 Напишите программу которая спрашивает у пользователя:
